@@ -14,50 +14,39 @@
     "prepublishOnly": "npm run lint && npm run build"
   },
   "dependencies": {
-    "@editorjs/editorjs": "^2.22.2",
-    "cropperjs": "^1.5.12",
+    "@editorjs/editorjs": "^2.28.0",
+    "cropperjs": "^1.6.0",
     "dot": "^2.0.0-beta.1",
-    "mini-ioc": "^1.0.0",
+    "mini-ioc": "^1.0.1",
     "mini-ioc-vue": "^0.3.0",
     "quill": "^1.3.7",
-    "vue": "^3.2.21",
-    "vue-router": "^4.0.11",
+    "vue": "^3.3.4",
+    "vue-router": "^4.2.4",
     "vuedraggable": "github:mrTimofey/vue.draggable.next"
   },
   "devDependencies": {
-    "@editorjs/delimiter": "^1.2.0",
-    "@editorjs/embed": "^2.5.0",
-    "@editorjs/footnotes": "^1.1.0",
-    "@editorjs/header": "^2.6.2",
-    "@editorjs/image": "^2.6.2",
-    "@editorjs/link": "^2.3.1",
-    "@editorjs/nested-list": "^1.0.2",
-    "@editorjs/quote": "^2.4.0",
-    "@editorjs/raw": "^2.3.0",
-    "@editorjs/table": "^2.0.1",
-    "@editorjs/underline": "^1.0.0",
-    "@types/node": "^18.11.9",
-    "@types/quill": "^2.0.9",
-    "@typescript-eslint/eslint-plugin": "^5.3.1",
-    "@typescript-eslint/parser": "^5.3.1",
+    "@types/node": "^20.6.0",
+    "@types/quill": "^2.0.11",
+    "@typescript-eslint/eslint-plugin": "^6.7.0",
+    "@typescript-eslint/parser": "^6.7.0",
     "@vitejs/plugin-vue": "^2.0.1",
-    "@volar/vue-language-plugin-pug": "^1.0.9",
-    "eslint": "^8.2.0",
+    "@vue/language-plugin-pug": "^1.8.11",
+    "eslint": "^8.49.0",
     "eslint-config-airbnb-base": "^15.0.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-filenames": "^1.3.2",
-    "eslint-plugin-import": "^2.24.2",
-    "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-vue": "^9.0.1",
-    "prettier": "^2.4.1",
+    "eslint-plugin-import": "^2.28.1",
+    "eslint-plugin-prettier": "^5.0.0",
+    "eslint-plugin-vue": "^9.17.0",
+    "prettier": "^3.0.3",
     "pug": "^3.0.2",
-    "stylus": "^0.59.0",
-    "typescript": "^4.4.2",
+    "stylus": "^0.60.0",
+    "typescript": "^5.2.2",
     "vite": "^2.6.13",
     "vite-plugin-dts": "1.0",
     "vite-plugin-html": "^3.0.3",
-    "vue-eslint-parser": "^9.0.1",
-    "vue-tsc": "^1.0.9"
+    "vue-eslint-parser": "^9.3.1",
+    "vue-tsc": "^1.8.11"
   },
   "engines": {
     "node": ">=16",
