@@ -44,8 +44,7 @@
 				)
 					template(
 						v-for="(_, slot) of slots"
-						#[slot]=`/* @ts-ignore */
-									scope`
+						#[slot]="scope"
 					)
 						slot(:name="slot" v-bind="scope")
 </template>
